refactor: prepare for noogle - monorepo
