+++ .gitignore
@@ -7,8 +7,8 @@ node_modules
 
 # testing
 coverage
-models/data/*
-!models/data/index.ts
+website/models/data/*
+!website/models/data/index.ts
 
 # nix
 .direnv/
